requirements.txt
scikit-learn
jupyter
langchain
langchain-openai
python-dotenv
faiss-cpu
langchain_community 
tiktoken
langchainhub
chromadb